+++ src/lib/data/fetchers/index.ts
@@ -58,3 +58,6 @@ export { CosmosFetcher, getCosmosFetcher } from './cosmos';
 
 // Hyperliquid fetchers
 export { HyperliquidFetcher, getHyperliquidFetcher } from './hyperliquid';
+
+// Kaspa fetchers
+export { KaspaFetcher, getKaspaFetcher } from './kaspa';